requirements.txt
streamlit
google-generativeai
Pillow
langchain
langchain-google-genai
pypdf
python-docx
openpyxl
faiss-cpu
